# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: TEMPO ADHD-friendly AI daily planner API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /tasks:
    get:
      operationId: listTasks
      tags: [tasks]
      summary: List all tasks
      parameters:
        - name: status
          in: query
          schema:
            type: string
            enum: [inbox, today, scheduled, done, cancelled]
        - name: projectId
          in: query
          schema:
            type: integer
        - name: folderId
          in: query
          schema:
            type: integer
      responses:
        "200":
          description: List of tasks
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Task"
    post:
      operationId: createTask
      tags: [tasks]
      summary: Create a new task
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateTaskBody"
      responses:
        "201":
          description: Created task
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Task"
  /tasks/{id}:
    get:
      operationId: getTask
      tags: [tasks]
      summary: Get a task by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Task details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Task"
        "404":
          description: Not found
    patch:
      operationId: updateTask
      tags: [tasks]
      summary: Update a task
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateTaskBody"
      responses:
        "200":
          description: Updated task
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Task"
        "404":
          description: Not found
    delete:
      operationId: deleteTask
      tags: [tasks]
      summary: Delete a task
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
        "404":
          description: Not found
  /notes:
    get:
      operationId: listNotes
      tags: [notes]
      summary: List all notes
      parameters:
        - name: projectId
          in: query
          schema:
            type: integer
        - name: folderId
          in: query
          schema:
            type: integer
      responses:
        "200":
          description: List of notes
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Note"
    post:
      operationId: createNote
      tags: [notes]
      summary: Create a new note
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateNoteBody"
      responses:
        "201":
          description: Created note
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Note"
  /notes/{id}:
    get:
      operationId: getNote
      tags: [notes]
      summary: Get a note by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Note details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Note"
        "404":
          description: Not found
    patch:
      operationId: updateNote
      tags: [notes]
      summary: Update a note
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateNoteBody"
      responses:
        "200":
          description: Updated note
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Note"
        "404":
          description: Not found
    delete:
      operationId: deleteNote
      tags: [notes]
      summary: Delete a note
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
        "404":
          description: Not found
  /projects:
    get:
      operationId: listProjects
      tags: [projects]
      summary: List all projects
      responses:
        "200":
          description: List of projects
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Project"
    post:
      operationId: createProject
      tags: [projects]
      summary: Create a new project
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateProjectBody"
      responses:
        "201":
          description: Created project
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Project"
  /projects/{id}:
    patch:
      operationId: updateProject
      tags: [projects]
      summary: Update a project
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateProjectBody"
      responses:
        "200":
          description: Updated project
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Project"
        "404":
          description: Not found
    delete:
      operationId: deleteProject
      tags: [projects]
      summary: Delete a project
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
        "404":
          description: Not found
...
  /tags:
    get:
      operationId: listTags
      tags: [tags]
      summary: List all tags
      responses:
        "200":
          description: List of tags
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Tag"
    post:
      operationId: createTag
      tags: [tags]
      summary: Create a new tag
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateTagBody"
      responses:
        "201":
          description: Created tag
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Tag"
  /tags/{id}:
    delete:
      operationId: deleteTag
      tags: [tags]
      summary: Delete a tag
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
        "404":
          description: Not found
...
  /preferences:
    get:
      operationId: getPreferences
      tags: [preferences]
      summary: Get user preferences
      responses:
        "200":
          description: User preferences
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PreferenceMemory"
    put:
      operationId: updatePreferences
      tags: [preferences]
      summary: Update user preferences
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdatePreferencesBody"
      responses:
        "200":
          description: Updated preferences
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PreferenceMemory"
...
  /onboarding:
    post:
      operationId: completeOnboarding
      tags: [onboarding]
      summary: Complete onboarding and save preferences
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/OnboardingBody"
      responses:
        "200":
          description: Onboarding completed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PreferenceMemory"
...
components:
  schemas:
...
    Task:
      type: object
      properties:
        id:
          type: integer
        title:
          type: string
        status:
          type: string
          enum: [inbox, today, scheduled, done, cancelled]
        priority:
          type: string
          enum: [high, medium, low]
        projectId:
          type: integer
          nullable: true
        folderId:
          type: integer
          nullable: true
        tags:
          type: array
          items:
            type: string
        dueDate:
          type: string
          nullable: true
        scheduledDate:
          type: string
          nullable: true
        estimatedMinutes:
          type: integer
          nullable: true
        notes:
          type: string
          nullable: true
        parentTaskId:
          type: integer
          nullable: true
        aiGenerated:
          type: boolean
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
      required:
        - id
        - title
        - status
        - priority
        - tags
        - aiGenerated
        - createdAt
        - updatedAt
    CreateTaskBody:
      type: object
      properties:
        title:
          type: string
        status:
          type: string
          enum: [inbox, today, scheduled, done, cancelled]
          default: inbox
        priority:
          type: string
          enum: [high, medium, low]
          default: medium
        projectId:
          type: integer
          nullable: true
        folderId:
          type: integer
          nullable: true
        tags:
          type: array
          items:
            type: string
        dueDate:
          type: string
          nullable: true
        scheduledDate:
          type: string
          nullable: true
        estimatedMinutes:
          type: integer
          nullable: true
        notes:
          type: string
          nullable: true
        parentTaskId:
          type: integer
          nullable: true
        aiGenerated:
          type: boolean
          default: false
      required:
        - title
    UpdateTaskBody:
      type: object
      properties:
        title:
          type: string
        status:
          type: string
          enum: [inbox, today, scheduled, done, cancelled]
        priority:
          type: string
          enum: [high, medium, low]
        projectId:
          type: integer
          nullable: true
        folderId:
          type: integer
          nullable: true
        tags:
          type: array
          items:
            type: string
        dueDate:
          type: string
          nullable: true
        scheduledDate:
          type: string
          nullable: true
        estimatedMinutes:
          type: integer
          nullable: true
        notes:
          type: string
          nullable: true
        parentTaskId:
          type: integer
          nullable: true
    Note:
      type: object
      properties:
        id:
          type: integer
        title:
          type: string
        content:
          type: string
        projectId:
          type: integer
          nullable: true
        folderId:
          type: integer
          nullable: true
        tags:
          type: array
          items:
            type: string
        templateType:
          type: string
          nullable: true
        isPinned:
          type: boolean
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
      required:
        - id
        - title
        - content
        - tags
        - isPinned
        - createdAt
        - updatedAt
    CreateNoteBody:
      type: object
      properties:
        title:
          type: string
        content:
          type: string
          default: ""
        projectId:
          type: integer
          nullable: true
        folderId:
          type: integer
          nullable: true
        tags:
          type: array
          items:
            type: string
        templateType:
          type: string
          nullable: true
        isPinned:
          type: boolean
          default: false
      required:
        - title
    UpdateNoteBody:
      type: object
      properties:
        title:
          type: string
        content:
          type: string
        projectId:
          type: integer
          nullable: true
        folderId:
          type: integer
          nullable: true
        tags:
          type: array
          items:
            type: string
        templateType:
          type: string
          nullable: true
        isPinned:
          type: boolean
    Project:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        description:
          type: string
          nullable: true
        folderId:
          type: integer
          nullable: true
        color:
          type: string
          nullable: true
        status:
          type: string
          enum: [active, archived]
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
      required:
        - id
        - name
        - status
        - createdAt
        - updatedAt
    CreateProjectBody:
      type: object
      properties:
        name:
          type: string
        description:
          type: string
          nullable: true
        folderId:
          type: integer
          nullable: true
        color:
          type: string
          nullable: true
        status:
          type: string
          enum: [active, archived]
          default: active
      required:
        - name
    UpdateProjectBody:
      type: object
      properties:
        name:
          type: string
        description:
          type: string
          nullable: true
        folderId:
          type: integer
          nullable: true
        color:
          type: string
          nullable: true
        status:
          type: string
          enum: [active, archived]
...
    Tag:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        color:
          type: string
          nullable: true
      required:
        - id
        - name
    CreateTagBody:
      type: object
      properties:
        name:
          type: string
        color:
          type: string
          nullable: true
      required:
        - name
...
    PreferenceMemory:
      type: object
      properties:
        id:
          type: integer
        wakeTime:
          type: string
        sleepTime:
          type: string
        energyPeaks:
          type: array
          items:
            type: string
        prepBufferMinutes:
          type: integer
        planningStyle:
          type: string
          enum: [morning, reactive, evening]
        adhdMode:
          type: boolean
        focusSessionMinutes:
          type: integer
        breakMinutes:
          type: integer
        onboardingComplete:
          type: boolean
        updatedAt:
          type: string
          format: date-time
      required:
        - id
        - wakeTime
        - sleepTime
        - energyPeaks
        - prepBufferMinutes
        - planningStyle
        - adhdMode
        - focusSessionMinutes
        - breakMinutes
        - onboardingComplete
        - updatedAt
    UpdatePreferencesBody:
      type: object
      properties:
        wakeTime:
          type: string
        sleepTime:
          type: string
        energyPeaks:
          type: array
          items:
            type: string
        prepBufferMinutes:
          type: integer
        planningStyle:
          type: string
          enum: [morning, reactive, evening]
        adhdMode:
          type: boolean
        focusSessionMinutes:
          type: integer
        breakMinutes:
          type: integer
        onboardingComplete:
          type: boolean
...
    OnboardingBody:
      type: object
      properties:
        adhdMode:
          type: boolean
        planningStyle:
          type: string
          enum: [morning, reactive, evening]
        wakeTime:
          type: string
        sleepTime:
          type: string
        energyPeaks:
          type: array
          items:
            type: string
        focusSessionMinutes:
          type: integer
        breakMinutes:
          type: integer
        prepBufferMinutes:
          type: integer
      required:
        - adhdMode
        - planningStyle
        - wakeTime
        - sleepTime
...