# lib/api-zod/src/generated/types/onboardingBody.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * TEMPO ADHD-friendly AI daily planner API
 * OpenAPI spec version: 0.1.0
 */
import type { OnboardingBodyPlanningStyle } from "./onboardingBodyPlanningStyle";

export interface OnboardingBody {
  adhdMode: boolean;
  planningStyle: OnboardingBodyPlanningStyle;
  wakeTime: string;
  sleepTime: string;
  energyPeaks?: string[];
  focusSessionMinutes?: number;
  breakMinutes?: number;
  prepBufferMinutes?: number;
}