# lib/api-zod/src/generated/types/updateTaskBody.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * TEMPO ADHD-friendly AI daily planner API
 * OpenAPI spec version: 0.1.0
 */
import type { UpdateTaskBodyPriority } from "./updateTaskBodyPriority";
import type { UpdateTaskBodyStatus } from "./updateTaskBodyStatus";

export interface UpdateTaskBody {
  title?: string;
  status?: UpdateTaskBodyStatus;
  priority?: UpdateTaskBodyPriority;
  projectId?: number | null;
  folderId?: number | null;
  tags?: string[];
  dueDate?: string | null;
  scheduledDate?: string | null;
  estimatedMinutes?: number | null;
  notes?: string | null;
  parentTaskId?: number | null;
}